moved the leaderboard types into the banter namespace
