+++ Runtime/Scripts/Models/BanterLeaderboard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5b697465cc22c241aacba3df6e85ead
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
